Moving a couple of types to @grafana/ui
