+++ packages/grafana-ui/src/types/index.ts
@@ -1,1 +1,3 @@
 export * from './series';
+export * from './time';
+export * from './panel';
