+++ public/app/types/panel.ts
@@ -1,35 +1,3 @@
-import { TimeSeries, TimeRange } from './series';
-import { LoadingState } from '@grafana/ui';
-
-export interface PanelProps<T = any> {
-  timeSeries: TimeSeries[];
-  timeRange: TimeRange;
-  loading: LoadingState;
-  options: T;
-  renderCounter: number;
-  width: number;
-  height: number;
-}
-
-export interface PanelOptionsProps<T = any> {
-  options: T;
-  onChange: (options: T) => void;
-}
-
-export interface PanelSize {
-  width: number;
-  height: number;
-}
-
-export interface PanelMenuItem {
-  type?: 'submenu' | 'divider';
-  text?: string;
-  iconClassName?: string;
-  onClick?: () => void;
-  shortcut?: string;
-  subMenu?: PanelMenuItem[];
-}
-
 export interface Threshold {
   index: number;
   value: number;
