+++ public/app/types/series.ts
@@ -1,68 +1,5 @@
-import { Moment } from 'moment';
 import { PluginMeta } from './plugins';
-
-export interface RawTimeRange {
-  from: Moment | string;
-  to: Moment | string;
-}
-
-export interface TimeRange {
-  from: Moment;
-  to: Moment;
-  raw: RawTimeRange;
-}
-
-export interface IntervalValues {
-  interval: string; // 10s,5m
-  intervalMs: number;
-}
-
-export type TimeSeriesValue = string | number | null;
-
-export type TimeSeriesPoints = TimeSeriesValue[][];
-
-export interface TimeSeries {
-  target: string;
-  datapoints: TimeSeriesPoints;
-  unit?: string;
-}
-
-/** View model projection of a time series */
-export interface TimeSeriesVM {
-  label: string;
-  color: string;
-  data: TimeSeriesValue[][];
-  stats: TimeSeriesStats;
-}
-
-export interface TimeSeriesStats {
-  total: number;
-  max: number;
-  min: number;
-  logmin: number;
-  avg: number | null;
-  current: number | null;
-  first: number | null;
-  delta: number;
-  diff: number | null;
-  range: number | null;
-  timeStep: number;
-  count: number;
-  allIsNull: boolean;
-  allIsZero: boolean;
-}
-
-export enum NullValueMode {
-  Null = 'null',
-  Ignore = 'connected',
-  AsZero = 'null as zero',
-}
-
-/** View model projection of many time series */
-export interface TimeSeriesVMs {
-  [index: number]: TimeSeriesVM;
-  length: number;
-}
+import { TimeSeries, TimeRange, RawTimeRange } from '@grafana/ui';
 
 export interface DataQueryResponse {
   data: TimeSeries[];
